[#173146890] align bonus available type with the remote data refactoring
