+++ ts/features/bonusVacanze/components/AvailableBonusItem.tsx
@@ -7,7 +7,7 @@ import IconFont from "../../../components/ui/IconFont";
 import { makeFontStyleObject } from "../../../theme/fonts";
 import customVariables from "../../../theme/variables";
 import { formatDateAsLocal } from "../../../utils/dates";
-import { BonusItem } from "../types/bonusList";
+import { BonusItem } from "../types/bonusesAvailable";
 
 type Props = {
   bonusItem: BonusItem;
