+++ ts/features/bonusVacanze/components/RequestBonus.tsx
@@ -7,7 +7,7 @@ import H5 from "../../../components/ui/H5";
 import I18n from "../../../i18n";
 import customVariables from "../../../theme/variables";
 import { maybeInnerProperty } from "../../../utils/options";
-import { BonusList } from "../types/bonusList";
+import { BonusList } from "../types/bonusesAvailable";
 import { BonusVacanze } from "../types/bonusVacanzeActivation";
 import { ID_BONUS_VACANZE_TYPE } from "../utils/bonus";
 import ActiveBonus from "./ActiveBonus";
