+++ ts/features/bonusVacanze/screens/AvailableBonusScreen.tsx
@@ -23,7 +23,7 @@ import AvailableBonusItem from "../components/AvailableBonusItem";
 import { availableBonusesLoad } from "../store/actions/bonusVacanze";
 import { availableBonusesSelector } from "../store/reducers/availableBonuses";
 import { bonusVacanzeActivationSelector } from "../store/reducers/bonusVacanzeActivation";
-import { BonusItem } from "../types/bonusList";
+import { BonusItem } from "../types/bonusesAvailable";
 import { BonusVacanze } from "../types/bonusVacanzeActivation";
 import { ID_BONUS_VACANZE_TYPE, isBonusActive } from "../utils/bonus";
