+++ ts/features/bonusVacanze/screens/BonusInformationScreen.tsx
@@ -14,7 +14,7 @@ import { navigateBack } from "../../../store/actions/navigation";
 import { ReduxProps } from "../../../store/actions/types";
 import themeVariables from "../../../theme/variables";
 import { FooterTwoButtons } from "../components/markdown/FooterTwoButtons";
-import { BonusItem } from "../types/bonusList";
+import { BonusItem } from "../types/bonusesAvailable";
 
 type NavigationParams = Readonly<{
   bonusItem: BonusItem;
